pyproject.toml
[project]
name = "fastapi-crud-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (==0.116.1)",
    "pyjwt (==2.10.1)",
    "python-dotenv (==1.1.1)",
    "returns (==0.26.0)",
    "sqlalchemy (==2.0.43)",
    "sqlmodel (==0.0.24)",
    "uvicorn (==0.35.0)",
    "asyncpg (==0.30.0)",
    "python-multipart (==0.0.20)",
    "alembic (==1.16.4)"
]

[tool.poetry]
packages = [{include = "fastapi_crud_demo", from = "src"}]


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
